// src/features/casino-discovery/schema/schema.ts
import { z } from 'zod';

export const CasinoSchema = z.object({
  state_abbreviation: z.string(),
  casinos: z.array(
    z.object({
      casino_name: z.string(),
      website: z.string(),
      license_status: z.enum(['active', 'pending', 'unknown']),
      source_url: z.string(),
    })
  ),
});

export const DiscoverCasinoSchema = z.object({
  discoverCasino: z.array(CasinoSchema),
});

export type DiscoverCasino = z.infer<typeof CasinoSchema>;